refactor(signals): extract detectors-phase + hotspot-detection helpers
The Story #1770 wiring landed inline in post-merge-pipeline.js
(detectorsPhase) and progress-reporter.js (runHotspotDetection); both
helpers were big enough to push their host modules below the
maintainability baseline. Move them into sibling modules
(lib/orchestration/detectors-phase.js,
lib/orchestration/epic-runner/hotspot-detection.js) so each call site
stays focused: the post-merge sequencer holds only phase orchestration
and the progress reporter holds only the periodic-render surface.

Re-export both functions from their original locations for backwards
compatibility — existing callers (and the tests added in #1779/#1780)
continue to import unchanged.

baseline-refresh: ratchet maintainability for the refactor — the new
sibling modules score 94.8 / 91.3 respectively; the parent files
recover to 86.82 / 91.549, both within 0.5 tolerance of the prior
baseline.
