+++ baselines/maintainability.json
@@ -145,6 +145,7 @@
   ".agents/scripts/lib/orchestration/context-hydration-engine.js": 92.569,
   ".agents/scripts/lib/orchestration/context.js": 102.013,
   ".agents/scripts/lib/orchestration/dependency-analyzer.js": 87.952,
+  ".agents/scripts/lib/orchestration/detectors-phase.js": 94.803,
   ".agents/scripts/lib/orchestration/dispatch-engine.js": 98.293,
   ".agents/scripts/lib/orchestration/dispatch-pipeline.js": 93.83,
   ".agents/scripts/lib/orchestration/doc-reader.js": 102.728,
@@ -158,10 +159,11 @@
   ".agents/scripts/lib/orchestration/epic-runner/commit-assertion.js": 97.501,
   ".agents/scripts/lib/orchestration/epic-runner/deliver-phases.js": 118.393,
   ".agents/scripts/lib/orchestration/epic-runner/factory.js": 96.908,
+  ".agents/scripts/lib/orchestration/epic-runner/hotspot-detection.js": 91.294,
   ".agents/scripts/lib/orchestration/epic-runner/phases/build-wave-dag.js": 105.841,
   ".agents/scripts/lib/orchestration/epic-runner/phases/iterate-waves.js": 91.835,
   ".agents/scripts/lib/orchestration/epic-runner/phases/snapshot.js": 127.788,
-  ".agents/scripts/lib/orchestration/epic-runner/progress-reporter.js": 90.613,
+  ".agents/scripts/lib/orchestration/epic-runner/progress-reporter.js": 91.549,
   ".agents/scripts/lib/orchestration/epic-runner/progress-signals/crap-drift.js": 86.949,
   ".agents/scripts/lib/orchestration/epic-runner/progress-signals/maintainability-drift.js": 82.999,
   ".agents/scripts/lib/orchestration/epic-runner/progress-signals/stalled-worktree.js": 116.374,
@@ -187,7 +189,7 @@
   ".agents/scripts/lib/orchestration/plan-runner/worktree-sweep.js": 82.927,
   ".agents/scripts/lib/orchestration/planning-context-budget.js": 96.421,
   ".agents/scripts/lib/orchestration/planning-state-manager.js": 98.578,
-  ".agents/scripts/lib/orchestration/post-merge-pipeline.js": 84.444,
+  ".agents/scripts/lib/orchestration/post-merge-pipeline.js": 86.82,
   ".agents/scripts/lib/orchestration/reconciler.js": 107.628,
   ".agents/scripts/lib/orchestration/recut.js": 111.264,
   ".agents/scripts/lib/orchestration/retro-heuristics.js": 114.637,
